moves enabled ads checks to their own test file
